move MetaVJ struct and get_type_e method from type.h
